resume-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic>=0.84.0
fpdf2==2.8.1
jinja2==3.1.4
python-dotenv==1.0.1
python-multipart==0.0.12
pdfplumber==0.11.4
